Update Spanish translations
Hey guys,

Congratulations for this awesome project! I'm from a region at the south-east of Spain (Almería) where farming is the main way of living for people. It's known as "The orchard of Europe" because there are thousands of greenhouses. I just found watched a video of this project on Facebook and I was so impressed that I wanted to show it to some of my relatives which are farmers. I noticed that the app was not translated in Spanish (my relatives doesn't speak English) so I decided to go ahead and submit this Pull Request.

I will follow this project very close and I hope that can do more contributions. Do not hesitate to ping me if you guys need some more Spanish translations.

Cheers!